rename test tsBuildInfoFile & rename mock file name
