+++ packages/playground/src/demo/display/vm-extension/vm-extension-list-demo.tsx
@@ -5,7 +5,7 @@ import {
 } from "@batch/ui-react/lib/vm-extension/vm-extension-list";
 import { DemoPane } from "../../../layout/demo-pane";
 // import { VmExtensionDetailsPanel } from "@batch/ui-react";
-import { allExtItems } from "@batch/ui-react/lib/vm-extension/mock-vme-ext-items";
+import { allExtItems } from "@batch/ui-react/lib/vm-extension/mock-vm-ext-items";
 
 export const VmExtensionListDemo: React.FC = () => {
     // const [panelOpen, setPanelOpen] = React.useState<boolean>(false);
