+++ packages/react/src/vm-extension/__tests__/vm-extension-details.spec.tsx
@@ -1,5 +1,5 @@
 import { render } from "@testing-library/react";
-import { failedExtItem, succeededExtItem } from "../mock-vme-ext-items";
+import { failedExtItem, succeededExtItem } from "../mock-vm-ext-items";
 import * as React from "react";
 import { initMockBrowserEnvironment } from "@azure/bonito-ui";
 import { runAxe } from "@azure/bonito-ui/lib/test-util/a11y";
