+++ packages/react/src/vm-extension/__tests__/vm-extension-list.spec.tsx
@@ -8,7 +8,7 @@ import {
     failedExtItem,
     noProvisioningStateExtItem,
     succeededExtItem,
-} from "../mock-vme-ext-items";
+} from "../mock-vm-ext-items";
 import * as React from "react";
 import { initMockBrowserEnvironment } from "@azure/bonito-ui";
 import { runAxe } from "@azure/bonito-ui/lib/test-util/a11y";
