+++ web/config/tsconfig.test.json
@@ -1,7 +1,7 @@
 {
   "extends": "./tsconfig.common.json",
   "compilerOptions": {
-    "tsBuildInfoFile": "../build/cache/ts.build.buildinfo",
+    "tsBuildInfoFile": "../build/cache/ts.test.buildinfo",
     "allowJs": true
   },
   "include": ["../**/__mocks__/**/*", "../**/__tests__/**/*"]
